Avoid invalid items from being saved (#43)
When creating a new item (private) and passing invalid input (e.g., no speech
but with a valid image), we're creating an invalid item and saving it to db,
even though we return an error. Item does not end up referencing the user as
it's invalidity.

Submit button is actually being changed from 'create item' to 'update item'
after returned error. If we subsequently try to 'update' the item, we get
"undefined method `update' for nil:NilClass" as the user does not actually
have a reference to the item.

Also, besides the fact we're creating tangling items (invalid and no reference
to the user), associated images end up saved to the disk, including spoofed
images (malicious) or any other invalid type.